fixed an error when updating

+++ app/Http/Controllers/EmployeeController.php
@@ -13,22 +13,25 @@ use Illuminate\Support\Facades\Auth;
 
 class EmployeeController extends Controller
 {
-    protected function index() {
+    protected function index()
+    {
         return view('employee_information_system.employee_list', [
             'Lists' => Employee::all()
         ]);
     }
 
-    protected function show(Employee $list) {
+    protected function show(Employee $list)
+    {
         return view('employee_information_system.profile', [
             'list' => $list
         ]);
     }
-    protected function edit(Employee $list){
+    protected function edit(Employee $list)
+    {
         //$user_role=$list::find(auth()->id)->get();
         //dd(Auth::user()->roles);
         $list_roles = $list->roles;
-        
+
         $array = null;
 
         $user_roles = Auth::user()->roles->all();
@@ -43,10 +46,10 @@ class EmployeeController extends Controller
 
             $user_array[] = $role->role_id;
         }
-        $is_admin=false;
-        if($user_array){
+        $is_admin = false;
+        if ($user_array) {
             if (in_array(1, $user_array)) {
-                $is_admin=true;
+                $is_admin = true;
             }
         }
         return view('employee_information_system.edit', [
@@ -54,17 +57,18 @@ class EmployeeController extends Controller
             'Departments' => Department::all(),
             'Roles' => Role::all(),
             'Positions' => Position::all(),
-            'is_admin'=>$is_admin,
-            'user_roles'=>$array        
+            'is_admin' => $is_admin,
+            'user_roles' => $array
         ]);
     }
 
     // update employee
-    public function update(Employee $list){
+    public function update(Employee $list)
+    {
         //dd(request()->all());        
         $formFields = request()->validate([
-            'work_start'=>['required' , 'string'],
-            'work_end'=>['required', 'string'],
+            'work_start' => ['required', 'string'],
+            'work_end' => ['required', 'string'],
             'first_name' => ['required', 'string', 'max:255'],
             'last_name' => ['required', 'string', 'max:255'],
             'address' => ['required', 'string', 'max:255'],
@@ -72,22 +76,28 @@ class EmployeeController extends Controller
             'birthday' => ['required', 'date'],
             'email' => ['required', 'string', 'email', 'max:255'],
             'contact_number' => ['required', 'string'],
-            'position_id'=> ['string']
+            'position_id' => ['string']
         ]);
         //dd($list->id);
         $list->update($formFields);
         $roles = request()->roles;
-        EmployeeRole::where('employee_id', $list->id)->delete();
-        foreach($roles as $role){
-            if (count(EmployeeRole::where('employee_id', $list->id)->where('role_id', $role)->get())== 0) {
-                EmployeeRole::create(['employee_id' => $list->id, "role_id" => $role]);
+        // dd($roles);
+
+        if ($roles) {
+            EmployeeRole::where('employee_id', $list->id)->delete();
+            foreach ($roles as $role) {
+                if (count(EmployeeRole::where('employee_id', $list->id)->where('role_id', $role)->get()) == 0) {
+                    EmployeeRole::create(['employee_id' => $list->id, "role_id" => $role]);
+                }
             }
         }
+
         return redirect("/employee_information_system/profile/{$list->id}");
     }
 
     // delete employee
-    protected function delete(Employee $list) {
+    protected function delete(Employee $list)
+    {
 
         if ($list->role != "Admin" || $list->id == Auth::user()->id) {
             abort(403, 'Already Logged in!');
@@ -96,5 +106,4 @@ class EmployeeController extends Controller
         $list->delete();
         return back();
     }
-
 }
